Fix/test structure (#2)
* change test structure

* move test to tests, set variables in ZUC struct to public
